Reduced ruby magic in detail parsing. Discourse notifications detail messages compilation adjustment.
